* Update the Topology class implementation:   - The port IDs are type "integer" instead of "short" so it will be easier     to use it for more general purpose (e.g., topologies with lots of tunnels).   - Implemented methods to get/add/remove various topology elements:     switch, port, link   - Added state to keep track of the reversed unidirectional links, so they     can be properly removed if a switch or a port is removed. * Add a Topology instance inside the PathComputation class, and   populate it as appropriate with ENTRY_ADD/ENTRY_REMOVE topology events   from the datagrid.

+++ src/main/java/net/onrc/onos/ofcontroller/topology/ShortestPath.java
@@ -123,7 +123,7 @@ public class ShortestPath {
 	Port outPort;
 	for (Node.Link link: resultPath) {
 	    // Setup the outgoing port, and add the Flow Entry
-	    outPort = new Port(link.myPort);
+	    outPort = new Port((short)link.myPort);
 
 	    FlowEntry flowEntry = new FlowEntry();
 	    flowEntry.setDpid(new Dpid(link.me.nodeId));
@@ -132,7 +132,7 @@ public class ShortestPath {
 	    result_data_path.flowEntries().add(flowEntry);
 
 	    // Setup the next incoming port
-	    inPort = new Port(link.neighborPort);
+	    inPort = new Port((short)link.neighborPort);
 	}
 	if (resultPath.size() > 0) {
 	    // Add the last Flow Entry
